feat(acmeat.domains.order,acmeat,domains,deliverycompany): fixed order date and deliverycompany price resetted to 0.

+++ acmeat/acmeat-be/server/acmeat.domains.deliverycompany/acmeat.domains.deliverycompany.models/Models/DeliveryCompany.cs
@@ -10,7 +10,7 @@ public class DeliveryCompany
 
     [JsonConstructor]
     public DeliveryCompany(
-               int Id, string Address, int Price, bool Available, string Name
+               int Id, string Address, float Price, bool Available, string Name
                )
     {
         this.Id = Id;
@@ -33,7 +33,7 @@ public class DeliveryCompany
     public int Id { get; set; }
     public string Address { get; set; }
     
-    public int Price { get; set; }
+    public float Price { get; set; }
     public bool Available { get; set; }
     
     public string Name { get; set; }
@@ -44,7 +44,7 @@ public class DeliveryCompany
         deliverycompany.Id = this.Id;
         deliverycompany.Address = this.Address;
         deliverycompany.Available = this.Available;
-        deliverycompany.Price = this.Price;
+        deliverycompany.Price = (int)this.Price;
         deliverycompany.Name = this.Name;
         return deliverycompany;
     }
